Moviendo la necesidad de areas en los filtros de las propuestas ciudadanas a un proyecto afuera
