+++ votainteligente/votainteligente_settings.py
@@ -29,3 +29,4 @@ AUTHORITY_MODEL = "elections.models.Candidate"
 PROPOSALS_ADAPTER = "proposals_for_votainteligente.models.ProposalsAdapter"
 MAIL_SENDER_FUNCTION = 'votainteligente.send_mails.send_mail'
 MAIL_TO_STAFF_SENDER_FUNCTION = 'votainteligente.send_mails.send_mails_to_staff'
+DEFAULT_FITERSET_CLASS = 'proposals_for_votainteligente.filters.ProposalGeneratedAtFilter'
